+++ src/internal/validations.ts
@@ -1,3 +1,6 @@
+import { init } from './config';
+import type { IOptions } from './types';
+
 /**
  * Check whether a given numeric variable is within a specific range.
  * @param property The property to check
@@ -23,4 +26,27 @@ export const isPositive = (property: number): boolean => property >= 0;
  */
 export const isPositiveInteger = (property: number): boolean => {
 	return isPositive(property) && Number.isInteger(property);
+};
+
+/**
+ * Checks whether some invalid values are found in the options object passed by the user.
+ * @param options The options object specified by the user in the Svelte components
+ * @returns The final valid options object that can be used by the reveal function
+ */
+export const checkOptions = (options: IOptions): Required<IOptions> => {
+	const finalOptions = Object.assign({}, init, options);
+	const { threshold, opacity, delay, duration, blur, scale } = finalOptions;
+
+	if (
+		hasValidRange(threshold, 0, 1) &&
+		hasValidRange(opacity, 0, 1) &&
+		isPositive(delay) &&
+		isPositive(duration) &&
+		isPositive(blur) &&
+		isPositive(scale)
+	) {
+		return finalOptions;
+	} else {
+		throw new Error('Invalid options');
+	}
 };